consulting-BE/requirements.txt
# trimmed on Sep 3, 2025
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
passlib==1.7.4
python-jose==3.5.0
python-dotenv==1.1.1
openai==1.102.0
httpx==0.28.1
pydantic==2.11.7
